% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios735.R
\name{rowVars}
\alias{rowVars}
\title{Calculates the sample variance for each row in a matrix}
\usage{
rowVars(x)
}
\arguments{
\item{x}{a matrix with m rows and n columns}
}
\value{
m sample variances for m rows of the matrix
}
\description{
This function calculates the sample variance
for each row in a matrix.
}
\examples{

m <- 5
n <- 50
little.n <- n/2
set.seed(1)
x <- matrix(rnorm(m*n),nrow=m,ncol=n)
rowVars(x)

}
