% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{Calculate t-statistic}
\usage{
getT(x, f)
}
\arguments{
\item{x}{The data matrix containing values from the two groups}

\item{f}{A vector with "1" and "2" that tells you which columns in 'x'
corresponds to which group.}
}
\value{
t-statistic
}
\description{
Computes the two-sample t-tests (equal variance) between two groups.
}
\examples{
m <- 400
n <- 50
little.n <- n/2
set.seed(1)
x <- matrix(rnorm(m*n),nrow=m,ncol=n)
f <- gl(2,little.n)
getT(x, f)



}
