% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{Two-sample t-tests (equal variance) between two groups}
\usage{
getT(x, f)
}
\arguments{
\item{x}{a m*n matrix. Data in each row are considered separately.}

\item{f}{a vector with length of n. The elements can take value 1 or 2. The total number of 1 equals the total number of 2. If the ith element of the vector equals 1, the ith column of x belongs to group 1; otherwise, the ith column of x belongs to group 2.}
}
\value{
a vector of t statistics with length of m
}
\description{
This function computes m sets of two sample t-tests between two groups.
For each row, it compares data in the first group and the second group.
Each group are of the same sample size (half of the number of columns).
It assumes equal variance between the two groups.
}
\examples{
set.seed(1)
x <- matrix(rnorm(50*4),nrow=50,ncol=4)
f <- c(1, 1, 2, 2)
getT(x, f)

}
