% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{Vectorized t-Test}
\usage{
getT(x, f)
}
\arguments{
\item{x}{a matrix of numbers}

\item{f}{a vector of two factor levels}
}
\value{
a vector of t-statistics executed by rows of \code{x}.
}
\description{
\code{getT} takes a matrix and a vector which assigns levels to the columns
of the matrix and executes vectorized t-tests by rows
allowing a vector of t-statistics to be returned.
}
\details{
\code{f} must have two levels.
}
\examples{
set.seed(1)
getT(matrix(rnorm(20), nrow=2), gl(2, 5))

}
