% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{computes m two-sample t-tests (equal variance) between two groups}
\usage{
getT(x, f)
}
\arguments{
\item{x}{a m*n matrix of observations}

\item{f}{factors with 2 levels, each level has length n/2, can be generated using gl(2,n/2)}
}
\value{
a vector of t test statistics with length m
}
\description{
This function computes m (m > 1) two-sample t-tests (equal variance) between two groups
using simple R functions for vector/matrix calculations, which is very fast
}
\examples{

x <- matrix(rnorm(40*50),nrow=40,ncol=50)
f <- gl(2,25)
t_stat <- getT(x,f)

}
