% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{getT}
\alias{getT}
\title{Two-sample t-tests (equal variance) between two groups}
\usage{
getT(x, f)
}
\arguments{
\item{x}{a matrix with m rows and n columns}

\item{f}{a factor with 2 levels with a total length of n indicating
the group assignments for the n elements in a row}
}
\value{
\code{m} number of t-test statistics for \code{m} columns of the given matrix
}
\description{
This function calculates the two-sample t-test statistic (equal variance)
for each row in a matrix based on a given two-level factor.
}
\examples{

m <- 5
n <- 50
little.n <- n/2
set.seed(1)

x <- matrix(rnorm(m*n), nrow=m, ncol=n)
f <- gl(2, little.n)
getT(x, f)

}
