% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getT.R
\name{rowVars}
\alias{rowVars}
\title{Get a faster t-test}
\usage{
rowVars(x)
}
\arguments{
\item{x}{a matrix with m rows and n columns}
}
\value{
m sample variances for m rows of the matrix
}
\description{
This function provides a faster method to return the t-test results
}
\details{
The rowvars function calculates the variance for each row in a matrix
Based on the first function, the second function returns the t-test result between x and f
}
\examples{

m <- 5
n <- 50
little.n <- n/2
set.seed(1)
x <- matrix(rnorm(m*n),nrow=m,ncol=n)
rowVars(x)

}
